Use 0 as default $code
It is passed as parameter to the constructor of
`FacebookAuthenticationException`, which extends `FacebookSDKException`,
which extends `\Exception`.

The `code` parameter of the `Exception` constructor needs to be of type
`int` since PHP 8.0.

This happens, if there is a response, which does not contain
`{"code": 123}` in the body of the response. This can happen both if
there's a problem with Facebook and during tests.